+++ graphique/__init__.py
@@ -1,2 +1,2 @@
 from .interface import Dataset
-from .middleware import GraphQL, MetricsExtension, implement
+from .middleware import GraphQL, MetricsExtension, implement, typed